atomic: make ostree in iso reflect 2wk ref
Let's configure the ostree that is embedded in the ISO to have the 2wk
ref in it so people running kickstarts can not have to worry about the
updates ref and only need to know about the 2wk ref.

Basically this makes it so that people can run:

ostreesetup --osname="fedora-atomic" --remote="fedora-atomic" \
--url="file:////run/install/repo/content/repo" \
--ref="fedora-atomic/25/x86_64/docker-host" --nogpg

Without having to know that this embedded ostree was actually created
from fedora-atomic/25/x86_64/updates/docker-host.

Signed-off-by: Dusty Mabe <dusty@dustymabe.com>