atomic: set up for 2wk updates
This is inspired by 286c199 from the f25 branch. Basically we need
to be able to pull from one ref [1] at the time we create the ISO, but
configure it so that after an install using the ISO the user has an
install that is configured to pull from another ref [2].

This is how we do it in the branched composes (i.e. f25/f26). Even
though rawhide doesn't follow this updates/stable ref model it makes
sense for us to set up the templates like this for consistency. The
actual values passed in for ostree_install_ref and ostree_update_ref
can be the same.

[1] fedora/26/x86_64/updates/atomic-host
[2] fedora/26/x86_64/atomic-host

Signed-off-by: Dusty Mabe <dusty@dustymabe.com>
(cherry picked from commit 6faa74c21edfb7cd8fb97a979a3209cd96fb3420)