atomic: create update/install ostree repo vars
Use install_repo for during the creation of the ISO during the
run of a compose. Use update_repo for the configuration of the
remote in the %post of kickstart that runs during a ISO install.

For example, when we create the ISO during a pungi run we need
to pull the ostree commit from the repo that was just created.

During the compose we need to pull from:

    https://kojipkgs.fedoraproject.org/compose/atomic/rawhide/

But we need to configure the remote for users to point to:

    https://kojipkgs.fedoraproject.org/atomic/rawhide/

Which is where the commits get synced to after the compose.

Signed-off-by: Dusty Mabe <dusty@dustymabe.com>
(cherry picked from commit ae3e3d7ec9e364560d77fbb5e19ba68ebebd3240)